chat_game_backend/requirements.txt
fastapi==0.95.2

pydantic==1.10.13

uvicorn
fastapi-jwt-auth
PyJWT
passlib[bcrypt]
uvicorn[standard]
python-dotenv>=0.19
openai>=0.27
marshmallow>=3.0

redis>=5.0.0



asyncpg
databases[postgresql]
sqlalchemy


psycopg2-binary
